% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{apply.coeffs}
\alias{apply.coeffs}
\title{Apply coefficients}
\usage{
apply.coeffs(specdata, coeffs, var = "Gear", use.min = TRUE)
}
\arguments{
\item{specdata}{specdata}

\item{coeffs}{coeffs}
}
\description{
Apply coefficients
}
