% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{collapse.grid}
\alias{collapse.grid}
\title{Collapse annualy varying grids to a single grid}
\usage{
collapse.grid(grid)
}
\arguments{
\item{grid}{grid}
}
\description{
Collapse annualy varying grids to a single grid
}
