% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{est.cog}
\alias{est.cog}
\title{Estimate abudance-weighted centre of gravity}
\usage{
est.cog(fit, grid, mod = 1, cols = 1)
}
\arguments{
\item{fit}{fitted object}

\item{grid}{prediction grid}

\item{mod}{Model (Default: 1)}

\item{cols}{ages, if multiple ages were modelled. Default: 1}
}
\value{
data frame with cog in lat and lon for each year
}
\description{
Estimate abudance-weighted centre of gravity
}
