\name{fishdist-package}
\alias{fishdist-package}
\alias{fishdist}
\docType{package}
\title{
\packageTitle{fishdist}
}
\description{
\packageDescription{fishdist}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fishdist}
\packageIndices{fishdist}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{fishdist}

Maintainer: \packageMaintainer{fishdist}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
