% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{list.datras.variables.req}
\alias{list.datras.variables.req}
\alias{list.datras.variables.all}
\title{List all DATRAS variables}
\usage{
list.datras.variables.all()

list.datras.variables.req(swept.area.calculated = TRUE)
}
\arguments{
\item{swept.area.calculated}{Was swept area already calculated? Otherwise
include Include DATRAS variables that are required for the swept area
calculation. Default: TRUE}
}
\value{
List with all DATRAS variables for the HH and HL DATRAS data sets

List with the required DATRAS variables for the HH and HL DATRAS data sets
}
\description{
List all DATRAS variables

List required DATRAS variables
}
