% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{list.recom.models}
\alias{list.recom.models}
\title{List all model recommended structures}
\usage{
list.recom.models(
  specdata,
  use.ctime = TRUE,
  use.toy = TRUE,
  use.ctime.space = TRUE,
  use.toy.space = TRUE,
  use.swept.area = FALSE,
  use.depth = TRUE,
  use.sqrt.depth = FALSE,
  use.gear.as.fixed = FALSE,
  use.random.ship = FALSE,
  use.temp.lt.space = FALSE,
  dim.lat.lon = 256,
  dim.ctime = "nyears",
  dim.ctime.lat.lon = c("nyears", 10),
  dim.timeOfYear = 5,
  dim.timeOfYear.lat.lon = c(5, 30),
  dim.temp.lt.lat.lon = c(3, 50),
  single.mod = TRUE,
  offset.plus.5 = TRUE
)
}
\arguments{
\item{specdata}{Species data}

\item{use.toy}{Use time of year? (Caused problems for some species)}

\item{use.swept.area}{Use swept area? (Might not be available)}

\item{dim.lat.lon}{Dimensions of the basis for the lat - lon smooth term (1
number)}

\item{dim.ctime.lat.lon}{Dimensions of the basis for the ctime - lat - lon
smooth term (2 numbers)}

\item{dim.timeOfYear.lat.lon}{Dimensions of the basis for the timeOfYear -
lat - lon smooth term (2 numbers)}
}
\value{
List with all model structures
}
\description{
List all model recommended structures
}
