% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{list.recom.models2}
\alias{list.recom.models2}
\title{List all model recommended structures}
\usage{
list.recom.models2(
  specdata,
  dim.ToD = 5,
  dim.DoY = 10,
  dim.year = "nyears",
  dim.lat.lon = 256,
  dim.DoY.year = c(5, "nyears"),
  dim.lat.lon.ToD = c(30, 5),
  dim.lat.lon.DoY = c(30, 5),
  dim.lat.lon.year = c(30, "nyears")
)
}
\arguments{
\item{specdata}{Species data}

\item{dim.lat.lon}{Dimensions of the basis for the lat - lon smooth term (1
number)}

\item{use.toy}{Use time of year? (Caused problems for some species)}

\item{use.swept.area}{Use swept area? (Might not be available)}

\item{dim.ctime.lat.lon}{Dimensions of the basis for the ctime - lat - lon
smooth term (2 numbers)}

\item{dim.timeOfYear.lat.lon}{Dimensions of the basis for the timeOfYear -
lat - lon smooth term (2 numbers)}
}
\value{
List with all model structures
}
\description{
List all model recommended structures
}
