% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotfishdish.diag}
\alias{plotfishdish.diag}
\title{plot diagnostics}
\usage{
plotfishdish.diag(fit, mod = 1)
}
\arguments{
\item{fit}{fit}

\item{mod}{Select one of the models. Default: 1}
}
\value{
Nothing
}
\description{
plot diagnostics
}
