% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotfishdish.dist.cv}
\alias{plotfishdish.dist.cv}
\title{plot cv of fit}
\usage{
plotfishdish.dist.cv(
  fit,
  mod = NULL,
  year = NULL,
  grid.all = NULL,
  xlim = NULL,
  ylim = NULL,
  legend = TRUE,
  fixed.scale = TRUE,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  xaxt = NULL,
  yaxt = NULL,
  fixed.lims = TRUE,
  cols = cm.colors(7)[-1],
  breaks = NULL,
  min.val = NA,
  cut.cv = NULL,
  asp = 2,
  plot.land = TRUE,
  plot.obs = 2,
  average = FALSE,
  mfrow = NULL,
  legend.cex = 0.6,
  legend.ncol = 1
)
}
\arguments{
\item{fit}{fit}

\item{mod}{Select one of the models. Default: NULL}
}
\value{
Nothing
}
\description{
plot cv of fit
}
