% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotfishdish.dist}
\alias{plotfishdish.dist}
\title{plot fit}
\usage{
plotfishdish.dist(
  fit,
  mod = NULL,
  year = NULL,
  xlim = NULL,
  ylim = NULL,
  legend = TRUE,
  fixed.scale = TRUE,
  cols = rev(heat.colors(8)),
  cut.cv = NULL,
  asp = 2
)
}
\arguments{
\item{fit}{fit}

\item{mod}{Select one of the models. Default: NULL}
}
\value{
Nothing
}
\description{
plot fit
}
