% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotfishdish.dist.year}
\alias{plotfishdish.dist.year}
\title{plot fit}
\usage{
plotfishdish.dist.year(
  fit,
  mod = NULL,
  var.lim = FALSE,
  all.years = TRUE,
  sandeel_areas = NULL,
  tobisbanker_wgs84 = NULL,
  xlim = NULL,
  ylim = NULL,
  cex = 1
)
}
\arguments{
\item{fit}{fit}

\item{mod}{Select one of the models. Default: NULL}

\item{var.lim}{variable axis limits? Default: FALSE}

\item{all.years}{Plot all years? Otherwise plot 3 years: first or 1991, intermediate,
and last year. Default: TRUE}
}
\value{
Nothing
}
\description{
plot fit
}
