% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotfishdish.dist2}
\alias{plotfishdish.dist2}
\title{plot fit}
\usage{
plotfishdish.dist2(
  fit,
  mod = NULL,
  year = NULL,
  grid.all = NULL,
  xlim = NULL,
  ylim = NULL,
  legend = TRUE,
  fixed.scale = TRUE,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  xaxt = NULL,
  yaxt = NULL,
  fixed.lims = TRUE,
  cols = rev(heat.colors(8)),
  min.val = NA,
  cut.cv = NULL,
  asp = 2,
  cv.stripes = 1,
  plot.land = TRUE,
  plot.obs = FALSE,
  average = FALSE,
  mfrow = NULL,
  legend.cex = 0.6,
  legend.ncol = 1,
  stripes.lwd = 2,
  stripes.n = 200
)
}
\arguments{
\item{fit}{fit}

\item{mod}{Select one of the models. Default: NULL}
}
\value{
Nothing
}
\description{
plot fit
}
