% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotfishdish.gam.effects.gear}
\alias{plotfishdish.gam.effects.gear}
\title{plot gam effects}
\usage{
plotfishdish.gam.effects.gear(
  fit,
  mod = 1,
  xlim = NULL,
  ylim = NULL,
  CI = 0.95,
  var = "Gear",
  exp = TRUE,
  use.bio = FALSE
)
}
\arguments{
\item{fit}{fit}

\item{mod}{Select one of the models. Default: 1}
}
\value{
Nothing
}
\description{
plot gam effects
}
