% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotfishdish.gam.effects}
\alias{plotfishdish.gam.effects}
\title{plot gam effects}
\usage{
plotfishdish.gam.effects(fit, mod = 1, xlim = NULL, ylim = NULL)
}
\arguments{
\item{fit}{fit}

\item{mod}{Select one of the models. Default: 1}
}
\value{
Nothing
}
\description{
plot gam effects
}
