% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotfishdish.overlapp}
\alias{plotfishdish.overlapp}
\alias{plotfishdish.overlap}
\title{plot overlapp}
\usage{
plotfishdish.overlap(x, var = "Gear", xlim = NULL, ylim = NULL, combine = TRUE)
}
\arguments{
\item{x}{data}
}
\value{
Nothing
}
\description{
plot overlapp
}
