% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reduce.datras}
\alias{reduce.datras}
\title{reduce datras}
\usage{
reduce.datras(
  data,
  surveys = NULL,
  areas = NULL,
  species = NULL,
  aphias = NULL,
  use.ca = TRUE
)
}
\arguments{
\item{data}{Data}
}
\description{
reduce datras
}
