% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{remove.ones}
\alias{remove.ones}
\title{remove ones}
\usage{
remove.ones(
  data,
  response = "N",
  vars = c("Survey", "Gear", "ShipG"),
  min.obs = 1,
  cut.off = 0
)
}
\arguments{
\item{data}{Data}
}
\description{
remove ones
}
