% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fdist.fit}
\alias{summary.fdist.fit}
\title{Summary of Fit}
\usage{
\method{summary}{fdist.fit}(fit)
}
\arguments{
\item{Fit}{A fit of est.dist or est.dist.one function}
}
\description{
Summary of Fit
}
