% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.data.R
\name{download.data}
\alias{download.data}
\title{Download all surveys}
\usage{
download.data(
  first.year = 1967,
  last.year = 2023,
  surveys = "all",
  quarters = "all",
  aphiaID = "all",
  datasets = c("HH", "HL"),
  calc.swept.area = FALSE,
  datras.variables = list.datras.variables.req(swept.area.calculated = FALSE),
  reduce.file.size = TRUE,
  verbose = TRUE,
  save.files = TRUE,
  file.dir = "files"
)
}
\arguments{
\item{first.year}{First year (default: 1967)}

\item{last.year}{Last year (default: 2020)}

\item{surveys}{(default: "all")}

\item{quarters}{(default: "all")}

\item{aphiaID}{Downloads all species anyways, but early subsetting for
species optional and can help with memory issues.}

\item{datasets}{(default: "HH","HL")}

\item{calc.swept.area}{Calculate swept area?}

\item{datras.variables}{DATRAS variables to use (Only used if reduce.file.size = TRUE).}

\item{reduce.file.size}{Reduce file size}

\item{verbose}{Print information? Default: TRUE}

\item{save.files}{Save individual files for each year? Default: TRUE}
}
\value{
List containing hh and hl data sets.
}
\description{
Download all surveys
}
