% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.dist.R
\name{est.dist.one}
\alias{est.dist.one}
\title{Estimate distribution of one species}
\usage{
est.dist.one(
  specdata,
  mods = NULL,
  n.lon = 20,
  nBoot = 0,
  grid = NULL,
  use.bathy = FALSE,
  resolution = 20,
  max.dist = 3,
  var.grid = TRUE,
  use.obs.for.pred = TRUE,
  predfix = NULL,
  model.bio = FALSE,
  fam = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{specdata}{Data set of one species}

\item{mods}{NULL default}

\item{nBoot}{number of bootrstrapping samples}

\item{use.bathy}{Use bathymetry for distribution. Default: FALSE}

\item{resolution}{Resolution for bathymetry map. Default: 20}

\item{max.dist}{Maximum distance to sample locations for bathymetry map. Default: 3}

\item{var.grid}{Use variable grid by year. Default: TRUE.}

\item{use.obs.for.pred}{Logical, Only use observed hauls for prediction values? Default: TRUE}

\item{predfix}{Prediction data. Default: NULL}

\item{verbose}{Print stuff? Default: TRUE}
}
\value{
List containing model results
}
\description{
Estimate distribution of one species
}
