% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.gear.effect}
\alias{get.gear.effect}
\title{Subset}
\usage{
get.gear.effect(fit, mod = 1, CI = 0.95, var = "Gear", exp = TRUE)
}
\arguments{
\item{fit}{fit}

\item{mod}{1}

\item{CI}{0.95}
}
\description{
Subset
}
