% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.info.surveys}
\alias{get.info.surveys}
\title{Get some info about surveys}
\usage{
get.info.surveys(survey = NULL)
}
\arguments{
\item{survey}{optional vector or single survey}
}
\value{
Data frame with info
}
\description{
Get some info about surveys
}
