% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.data.R
\name{load.data}
\alias{load.data}
\title{Download all surveys}
\usage{
load.data(
  file.dir = "files",
  files = NULL,
  first.year = 1967,
  last.year = 2020,
  surveys = "all",
  quarters = "all",
  aphiaID = "all",
  datasets = c("HH", "HL"),
  calc.swept.area = TRUE,
  datras.variables = list.datras.variables.req(swept.area.calculated = !calc.swept.area),
  reduce.file.size = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{file.dir}{Directory with saved files, default: 'files'}

\item{files}{Optional list with names of files to load. If NULL all files in
folder 'file.dir' will be loaded}

\item{first.year}{First year (default: 1967)}

\item{last.year}{Last year (default: 2020)}

\item{surveys}{(default: "all")}

\item{quarters}{(default: "all")}

\item{aphiaID}{Downloads all species anyways, but early subsetting for
species optional and can help with memory issues.}

\item{datasets}{(default: "HH","HL")}

\item{calc.swept.area}{Calculate swept area?}

\item{datras.variables}{DATRAS variables to use (Only used if
reduce.file.size = TRUE).}

\item{reduce.file.size}{Reduce file size}

\item{verbose}{Print information? Default: TRUE}
}
\value{
List containing hh and hl data sets.
}
\description{
Download all surveys
}
