% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotfishdish.abun}
\alias{plotfishdish.abun}
\title{plot fit}
\usage{
plotfishdish.abun(
  fit,
  by.area = FALSE,
  by.eco = FALSE,
  ylab = "Abundance index",
  y.scale = 1,
  fixed.scale = FALSE
)
}
\arguments{
\item{fit}{fit}

\item{by}{by}
}
\value{
Nothing
}
\description{
plot fit
}
