% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.species.R
\name{prep.species}
\alias{prep.species}
\title{Prepare species}
\usage{
prep.species(
  data,
  aphiaID = NULL,
  use.gear.cat = TRUE,
  remove.fragmented.years = TRUE,
  filter.shipg = FALSE,
  filter.statrec = FALSE,
  min.depth = 10,
  max.depth = 1000,
  min.gears = 0,
  min.surveys = 0,
  min.hauls = 0,
  min.ship.gear = 0,
  verbose = TRUE
)
}
\arguments{
\item{data}{Data set}

\item{aphiaID}{required (single species or vector of species)}

\item{use.gear.cat}{= TRUE}

\item{remove.fragmented.years}{= TRUE}

\item{max.depth}{Maximum depth to inclde (default 1000m)}

\item{min.gears}{= 1}

\item{min.surveys}{= 1}

\item{min.hauls}{= 1}

\item{min.ship.gear}{= 1}

\item{verbose}{Print stuff Default: TRUE}
}
\value{
List containing hh and hl data sets.
}
\description{
Prepare species
}
